:truck: Extract status queries to own app
